Plugins: categorize linux.hidden_modules as a malware plugin
